Add channel retries support
We also add helpers
smb_has_multiple_channels() and
smb_get_next_connection().

The timeout values are currently hardcoded at
10 seconds for the first timeout value and
2 seconds for each subsequent retry on the new channel.
We continue retrying channels until we have retried the break on all
available channels. An oplock break on any of the xconns connected to
the client will complete the request.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>